provide macros-test crate to ensure tests work without dependencies in namespace
