+++ Cargo.toml
@@ -8,6 +8,7 @@ members = [
     "cryptography",
     "deployer",
     "macros",
+    "macros-test",
     "p2p",
     "proc-macros",
     "resolver",
@@ -51,6 +52,7 @@ commonware-consensus = { version = "0.0.62", path = "consensus" }
 commonware-cryptography = { version = "0.0.62", path = "cryptography", default-features = false }
 commonware-deployer = { version = "0.0.62", path = "deployer", default-features = false }
 commonware-macros = { version = "0.0.62", path = "macros" }
+commonware-macros-test = { version = "0.0.62", path = "macros-test" }
 commonware-p2p = { version = "0.0.62", path = "p2p" }
 commonware-proc-macros = { version = "0.0.62", path = "proc-macros" }
 commonware-resolver = { version = "0.0.62", path = "resolver" }
