+++ macros-test/src/lib.rs
@@ -1,0 +1,2 @@
+//! This crate is empty. It's only purpose is to assert that proc macros
+//! compile and work without having their dependencies in the namespace.
